Fix segmentation fault caused by hls-prepare for some files.
As discussed in github issue #11, some input files cause segmentation
faults during hls-prepare. In these cases the reason is that we access
data after the end of the audio master file vector.

This commit fixes the problem by appending zeros if the audio master is
too short to provide the context for a segment. This also allows us to
unconditionally using 3 seconds of context after every segment (even
after the last segment).

Signed-off-by: Stefan Westerfeld <stefan@space.twc.de>